PR testsuite/31589         * gcc.dg/vect/vect-iv-9.c: Added vect_int_mult target keyword to         dg-final test.         * gcc.dg/vect/vect-reduc-dot-u16b.c: Added vect_pack_trunc target         keyword to dg-final test.         * gcc.dg/vect/vect-iv-4.c: Likewise.         * gcc.dg/vect/vect-widen-mult-u16.c: Likewise.         * gcc.dg/vect/pr30771.c: Added vect_unapck target keyword to dg-final         test.         * gcc.dg/vect/vect-reduc-dot-u16a.c: Change variable type to avoid a         cast.         * gcc.dg/vect/no-section-anchors-vect-69.c: xfail on is64.         * lib/target-supports.exp         (check_effective_target_vect_widen_sum_hi_to_si): Added ia64.         (check_effective_target_vect_widen_sum_qi_to_hi): Added ia64.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@124315 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/testsuite/gcc.dg/vect/vect-reduc-dot-u16a.c
@@ -30,7 +30,7 @@ foo1(int len) {
 int main (void)
 {
   unsigned int dot1;
-  int i;
+  unsigned short i;
 
   check_vect ();
